AI: ranged troops should move when blocked by a strong unit (#2915)
close #2747

+++ src/fheroes2/battle/battle_board.cpp
@@ -163,7 +163,7 @@ void Battle::Board::SetEnemyQuality( const Unit & unit )
     }
 }
 
-s32 Battle::Board::GetDistance( s32 index1, s32 index2 )
+uint32_t Battle::Board::GetDistance( s32 index1, s32 index2 )
 {
     if ( isValidIndex( index1 ) && isValidIndex( index2 ) ) {
         const int dx = std::abs( ( index1 % ARENAW ) - ( index2 % ARENAW ) );
@@ -171,7 +171,7 @@ s32 Battle::Board::GetDistance( s32 index1, s32 index2 )
         const int roundingUp = index1 / ARENAW % 2;
 
         // hexagonal grid: you only move half as much on X axis when diagonal!
-        return dy + std::max( dx - ( dy + roundingUp ) / 2, 0 );
+        return static_cast<uint32_t>( dy + std::max( dx - ( dy + roundingUp ) / 2, 0 ) );
     }
 
     return 0;
@@ -493,7 +493,7 @@ Battle::Indexes Battle::Board::GetPassableQualityPositions( const Unit & b )
 
 std::vector<Battle::Unit *> Battle::Board::GetNearestTroops( const Unit * startUnit, const std::vector<Battle::Unit *> & blackList )
 {
-    std::vector<std::pair<Battle::Unit *, int32_t> > foundUnits;
+    std::vector<std::pair<Battle::Unit *, uint32_t> > foundUnits;
 
     for ( Cell & cell : *this ) {
         Unit * cellUnit = cell.GetUnit();
@@ -508,7 +508,7 @@ std::vector<Battle::Unit *> Battle::Board::GetNearestTroops( const Unit * startU
     }
 
     std::sort( foundUnits.begin(), foundUnits.end(),
-               []( const std::pair<Battle::Unit *, int32_t> & first, const std::pair<Battle::Unit *, int32_t> & second ) { return first.second < second.second; } );
+               []( const std::pair<Battle::Unit *, uint32_t> & first, const std::pair<Battle::Unit *, uint32_t> & second ) { return first.second < second.second; } );
 
     std::vector<Battle::Unit *> units;
     units.reserve( foundUnits.size() );
